update tests, and default settings
